Move organisation tests out of index to new file
